fix stepper navigation
