+++ src/components/core/wizard/desktop-wizard/__tests__/step-navigation.test.tsx
@@ -1,0 +1,36 @@
+import { render, screen } from 'test-setup';
+import { StepNavigationProps } from '@core/wizard/types';
+import StepNavigation from '../step-navigation';
+
+describe('StepNavigation Component', () => {
+    const props: StepNavigationProps = {
+        steps: [
+            {
+                title: 'Product',
+            },
+            {
+                title: 'App',
+            },
+            {
+                title: 'Complete',
+            },
+        ],
+        dark: false,
+        current_step_index: 0,
+        complete_steps_indexes: [],
+        onClick: jest.fn(),
+    };
+
+    it('StepNavigation renders properly', () => {
+        render(<StepNavigation {...props} />);
+
+        const step_navigation = screen.getByTestId('step-navigation');
+        const steps = screen.getAllByTestId('step-item');
+
+        expect(step_navigation).toBeInTheDocument();
+        expect(steps.length).toBe(3);
+        expect(steps[0]).toHaveTextContent(/product/i);
+        expect(steps[1]).toHaveTextContent(/app/i);
+        expect(steps[2]).toHaveTextContent(/complete/i);
+    });
+});
